Print stacktrace of exporter exception when exporter fails
